@@ -17,7 +17,8 @@ import os
 import re
 import subprocess
 import sys
-from typing import Any, Callable
+from collections.abc import Callable
+from typing import Any
 
 
 def get_keywords() -> dict[str, str]:
