@@ -48,8 +48,8 @@ arise, and the order we would like to be determined.
 """
 import copy
 from collections import defaultdict, deque, namedtuple
-from collections.abc import Iterable, Mapping, MutableMapping
-from typing import Any, Callable, Literal, NamedTuple, overload
+from collections.abc import Callable, Iterable, Mapping, MutableMapping
+from typing import Any, Literal, NamedTuple, overload
 
 from dask.core import get_dependencies, get_deps, getcycle, istask, reverse_dict
 from dask.typing import Key
