@@ -22,9 +22,12 @@ if NUMPY_GE_200:
     from numpy.exceptions import AxisError, ComplexWarning  # noqa: F401
     from numpy.lib.array_utils import normalize_axis_index, normalize_axis_tuple
 else:
-    from numpy import AxisError, ComplexWarning  # noqa: F401
-    from numpy.core.numeric import normalize_axis_index  # type: ignore[attr-defined]
-    from numpy.core.numeric import normalize_axis_tuple  # type: ignore[attr-defined]
+    from numpy import (  # type: ignore[no-redef, attr-defined] # noqa: F401
+        AxisError,
+        ComplexWarning,
+    )
+    from numpy.core.numeric import normalize_axis_index  # type: ignore[no-redef]
+    from numpy.core.numeric import normalize_axis_tuple  # type: ignore[no-redef]
 
 
 # Taken from scikit-learn:
